requirements.txt
fastapi
uvicorn
requests
psycopg2-binary
pytz
pydantic
python-multipart
python-jose[cryptography]
